Add `api` package, move GitHub client
