+++ apps/backend/package.json
@@ -19,6 +19,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@repo/api": "workspace:*",
     "@repo/db": "workspace:*",
     "@types/fs-extra": "^11.0.1",
     "cleye": "^1.3.2",
