+++ package.json
@@ -14,6 +14,7 @@
     "@ianvs/prettier-plugin-sort-imports": "^4.3.1",
     "dotenv-cli": "^7.4.2",
     "prettier": "^3.3.3",
-    "tsx": "^4.16.2"
+    "tsx": "^4.16.2",
+    "typescript": "^5.5.4"
   }
 }
