+++ packages/api/README.md
@@ -1,0 +1,3 @@
+# Best of JS `api` package
+
+Shared services called by all applications
